refactor, feat : user-model -> domain,
- common에 Tech_part 추가 (디스코드에 insert문 공지)
- User/model -> User/domain으로 이름 변경
