+++ src/main/java/com/potential_radar/PR/user/domain/ExperienceRange.java
@@ -1,0 +1,3 @@
+package com.potential_radar.PR.user.domain;
+
+public enum ExperienceRange { FRESHER, LT_1, Y1_3, Y5_10, GE_10, ETC }
