+++ src/main/java/com/potential_radar/PR/user/domain/Provider.java
@@ -1,3 +1,3 @@
-package com.potential_radar.PR.user.model;
+package com.potential_radar.PR.user.domain;
 
 public enum Provider { EMAIL, GOOGLE, KAKAO }
