+++ src/main/java/com/potential_radar/PR/user/domain/User.java
@@ -1,4 +1,4 @@
-package com.potential_radar.PR.user.model;
+package com.potential_radar.PR.user.domain;
 
 import jakarta.persistence.*;
 import lombok.Builder;
