+++ src/main/java/com/potential_radar/PR/user/dto/UserInfoResponse.java
@@ -1,6 +1,6 @@
 package com.potential_radar.PR.user.dto;
 
-import com.potential_radar.PR.user.model.User;
+import com.potential_radar.PR.user.domain.User;
 
 public record UserInfoResponse(
         Long id,
