--- src/main/java/com/potential_radar/PR/user/model/ExperienceRange.java
@@ -1,3 +1,0 @@
-package com.potential_radar.PR.user.model;
-
-public enum ExperienceRange { FRESHER, LT_1, Y1_3, Y5_10, GE_10 }
