+++ src/main/java/com/potential_radar/PR/user/repository/TechPartRepository.java
@@ -1,6 +1,6 @@
 package com.potential_radar.PR.user.repository;
 
-import com.potential_radar.PR.user.model.TechPart;
+import com.potential_radar.PR.user.domain.TechPart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TechPartRepository extends JpaRepository<TechPart, Long> {
