+++ src/main/java/com/potential_radar/PR/user/repository/UserProfileRepository.java
@@ -1,7 +1,7 @@
 package com.potential_radar.PR.user.repository;
 
-import com.potential_radar.PR.user.model.User;
-import com.potential_radar.PR.user.model.UserProfile;
+import com.potential_radar.PR.user.domain.User;
+import com.potential_radar.PR.user.domain.UserProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
