+++ src/main/java/com/potential_radar/PR/user/service/RefreshTokenService.java
@@ -1,9 +1,8 @@
 package com.potential_radar.PR.user.service;
 
 import com.potential_radar.PR.config.jwt.JwtProperties;
-import com.potential_radar.PR.config.jwt.TokenProvider;
-import com.potential_radar.PR.user.model.RefreshToken;
-import com.potential_radar.PR.user.model.User;
+import com.potential_radar.PR.user.domain.RefreshToken;
+import com.potential_radar.PR.user.domain.User;
 import com.potential_radar.PR.user.repository.RefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
